+++ app/serializers/pathfinder2/character_serializer.rb
@@ -8,12 +8,13 @@ module Pathfinder2
                :skills, :created_at, :subclasses, :background, :saving_throws_value, :saving_throws, :dying_condition_value,
                :avatar, :weapon_skills, :armor_skills, :coins, :load, :armor_class, :speed, :perception, :conditions,
                :ability_boosts_v2, :skill_boosts, :attacks, :provider, :lores, :selected_features, :spells_info, :class_dc,
-               :spell_attack, :spell_dc, :spent_spell_slots, :formatted_static_spells, :modified_abilities, :spell_list
+               :spell_attack, :spell_dc, :spent_spell_slots, :formatted_static_spells, :modified_abilities, :spell_list,
+               :can_have_pet, :can_have_familiar
 
     delegate :health, :abilities, :skills, :subclasses, :saving_throws_value, :dying_condition_value, :weapon_skills,
              :money, :lores, :armor_skills, :coins, :load, :armor_class, :speed, :perception, :conditions, :attacks,
              :features, :spells_info, :class_dc, :spell_attack, :spell_dc, :formatted_static_spells, :modified_abilities,
-             to: :decorator
+             :can_have_pet, :can_have_familiar, to: :decorator
     delegate :name, :id, :data, :created_at, to: :object
     delegate :level, :ability_boosts_v2, :skill_boosts, :selected_features, :race, :subrace, :main_class, :classes, :languages,
              :background, :saving_throws, :spent_spell_slots, :spell_list, to: :data